refactor: split mood and custom message stores
