+++ client/src/lib/mood-store.ts
@@ -1,0 +1,114 @@
+import { MoodEntry, InsertMoodEntry } from "@shared/schema";
+import { nanoid } from "nanoid";
+import { customMessageStore } from "./custom-message-store";
+
+const MOOD_STORAGE_KEY = "cozy-critter-moods";
+
+export class MoodStore {
+  public saveMoodEntry(entry: InsertMoodEntry): MoodEntry {
+    const moodEntry: MoodEntry = {
+      ...entry,
+      id: nanoid(),
+    };
+
+    const existingMoods = this.getAllMoodEntries();
+    const updatedMoods = [moodEntry, ...existingMoods];
+
+    localStorage.setItem(MOOD_STORAGE_KEY, JSON.stringify(updatedMoods));
+    return moodEntry;
+  }
+
+  public getAllMoodEntries(): MoodEntry[] {
+    try {
+      const stored = localStorage.getItem(MOOD_STORAGE_KEY);
+      if (!stored) return [];
+
+      const parsed = JSON.parse(stored);
+      return Array.isArray(parsed) ? parsed : [];
+    } catch (error) {
+      console.error("Error reading mood entries from localStorage:", error);
+      return [];
+    }
+  }
+
+  public getMoodEntriesByDateRange(startDate: Date, endDate: Date): MoodEntry[] {
+    const allMoods = this.getAllMoodEntries();
+    return allMoods.filter(
+      (mood) =>
+        mood.timestamp >= startDate.getTime() &&
+        mood.timestamp <= endDate.getTime()
+    );
+  }
+
+  public getRecentMoodEntries(limit: number = 10): MoodEntry[] {
+    const allMoods = this.getAllMoodEntries();
+    return allMoods.slice(0, limit);
+  }
+
+  public deleteMoodEntry(id: string): boolean {
+    try {
+      const existingMoods = this.getAllMoodEntries();
+      const updatedMoods = existingMoods.filter(mood => mood.id !== id);
+
+      if (existingMoods.length === updatedMoods.length) {
+        // No entry was found with the given id
+        return false;
+      }
+
+      localStorage.setItem(MOOD_STORAGE_KEY, JSON.stringify(updatedMoods));
+      return true;
+    } catch (error) {
+      console.error("Error deleting mood entry from localStorage:", error);
+      return false;
+    }
+  }
+
+  public clearAllMoodEntries(): void {
+    localStorage.removeItem(MOOD_STORAGE_KEY);
+  }
+
+  public clearAllData(): void {
+    this.clearAllMoodEntries();
+    customMessageStore.clearAllCustomMessages();
+    // Also clear theme preference to fully reset
+    localStorage.removeItem("cozy-critter-theme");
+  }
+
+  public getDataSummary(): {
+    moodEntries: number;
+    customMessages: number;
+    storageUsed: string;
+    oldestEntry?: Date;
+    newestEntry?: Date;
+  } {
+    const moods = this.getAllMoodEntries();
+    const messages = customMessageStore.getAllCustomMessages();
+
+    // Calculate approximate storage usage
+    const moodData = localStorage.getItem(MOOD_STORAGE_KEY) || '';
+    const messageData = localStorage.getItem("cozy-critter-custom-messages") || '';
+    const themeData = localStorage.getItem("cozy-critter-theme") || '';
+    const totalBytes = moodData.length + messageData.length + themeData.length;
+    const storageUsed = totalBytes < 1024 ? `${totalBytes} bytes` : `${(totalBytes / 1024).toFixed(1)} KB`;
+
+    let oldestEntry: Date | undefined;
+    let newestEntry: Date | undefined;
+
+    if (moods.length > 0) {
+      const timestamps = moods.map(m => m.timestamp).sort((a, b) => a - b);
+      oldestEntry = new Date(timestamps[0]);
+      newestEntry = new Date(timestamps[timestamps.length - 1]);
+    }
+
+    return {
+      moodEntries: moods.length,
+      customMessages: messages.length,
+      storageUsed,
+      oldestEntry,
+      newestEntry,
+    };
+  }
+}
+
+export const moodStore = new MoodStore();
+
